requirements.txt
openai-whisper==20250625
pyaudio==0.2.14
fastapi==0.116.0
websockets==15.0.1
uvicorn==0.35.0
